Deleted duplicate wizard inspiration folder. Added Zone Compiler to dev/tools: Converts text files into usable Wizmatch Zone files (.wzz).

+++ source/Actor.cpp
@@ -1,6 +1,6 @@
 #include "Actor.h"
 
-int16_t Actor::actorNumIncrementor = 0;
+int32_t Actor::actorNumIncrementor = 0;
 
 void Actor::ApplyVelocity( const float deltaTime )
 {
@@ -18,7 +18,7 @@ void Actor::SetRotationY( const float &newY )
     rotation = frmr::Vec2f( rotation.GetX(), newY );
 }
 
-int16_t Actor::GetActorNum() const
+int32_t Actor::GetActorNum() const
 {
     return actorNum;
 }
